feat: refactoring employee

+++ internal/employees/domain/employee_domain.go
@@ -4,16 +4,16 @@ import "context"
 
 type EmployeeRepository interface {
 	GetAll(ctx context.Context) (*[]Employee, error)
-	GetById(ctx context.Context, id int) (*Employee, error)
+	GetById(ctx context.Context, id int64) (*Employee, error)
 	Create(ctx context.Context, employee *Employee) (*Employee, error)
 	Update(ctx context.Context, employee *Employee) (*Employee, error)
-	Delete(ctx context.Context, id int) error
+	Delete(ctx context.Context, id int64) error
 }
 
 type EmployeeService interface {
 	GetAll(ctx context.Context) (*[]Employee, error)
-	GetById(ctx context.Context, id int) (*Employee, error)
+	GetById(ctx context.Context, id int64) (*Employee, error)
 	Create(ctx context.Context, employee *Employee) (*Employee, error)
 	Update(ctx context.Context, employee *Employee) (*Employee, error)
-	Delete(ctx context.Context, id int) error
+	Delete(ctx context.Context, id int64) error
 }
